Renamed splunk.deploy module to splunk.reconcile
