+++ pkg/controller/spark/spark_controller.go
@@ -29,7 +29,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/source"
 
 	enterprisev1 "github.com/splunk/splunk-operator/pkg/apis/enterprise/v1alpha2"
-	"github.com/splunk/splunk-operator/pkg/splunk/deploy"
+	splunkreconcile "github.com/splunk/splunk-operator/pkg/splunk/reconcile"
 )
 
 var log = logf.Log.WithName("controller_spark")
@@ -115,7 +115,7 @@ func (r *ReconcileSpark) Reconcile(request reconcile.Request) (reconcile.Result,
 	instance.TypeMeta.APIVersion = "enterprise.splunk.com/v1alpha2"
 	instance.TypeMeta.Kind = "Spark"
 
-	err = deploy.ReconcileSpark(r.client, instance)
+	err = splunkreconcile.ReconcileSpark(r.client, instance)
 	if err != nil {
 		return reconcile.Result{}, err
 	}
