+++ pkg/splunk/reconcile/deployment.go
