+++ pkg/splunk/reconcile/deployment_test.go
