+++ pkg/splunk/reconcile/indexer.go
