+++ pkg/splunk/reconcile/indexer_test.go
