+++ pkg/splunk/reconcile/spark.go
