+++ pkg/splunk/reconcile/spark_test.go
