+++ pkg/splunk/reconcile/splunkenterprise_test.go
