+++ pkg/splunk/reconcile/standalone.go
