+++ pkg/splunk/reconcile/standalone_test.go
