+++ pkg/splunk/reconcile/util.go
